+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Image Annotation Tool
 
+## 🏷 1.1.1 &#8239;&#8239;&#8239;&#8239; 🗓 2021-08-24
+
+### Fixes and changes
+* The code was completely refactored and cleaned up.
+* When user clicks with right click when using zoom tool the image is now zoomed-out.
+* Icons for polygon (previously path) tool and rectangle drawing tools were updated.
+* Fixed issue where theme was not applied to the code-block editor.
+
 ## 🏷 1.1.0 &#8239;&#8239;&#8239;&#8239; 🗓 2021-08-24
 
 ### New Features